# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: dnd-app-sim API
  version: 0.1.0-draft
  summary: REST replacement for the dnd-app `sim/` CLI, plus the new encounter simulator.
  description: |
    Contract-first design for the Spring Boot port of the build-optimization simulator.
    See `docs/api/README.md` for conventions (determinism, seeds, common random numbers,
    jobs, limits, errors) and the CLI-to-REST mapping.

    Status: DRAFT for review. Limits quoted in descriptions are proposed defaults
    (`sim.limits.*`), not measured values.
servers:
  - url: http://localhost:8080
    description: Local development (`local` profile)

# Every operation needs an API key except /actuator/health and /actuator/info (see docs/api/README.md, Security).
# The `local` profile switches authentication off.
security:
  - ApiKeyHeader: []
  - BearerKey: []
...
paths:
...
  /api/v1/jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Jobs]
      operationId: getJob
      summary: Job status and progress.
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: The job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/Busy' }

    delete:
      tags: [Jobs]
      operationId: cancelJob
      summary: Cancel a queued or running job.
      description: 'Cooperative: the job stops at the next generation boundary.'
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: The job after the cancel request (status `cancelled` once honored).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The job already finished.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/Busy' }
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerKey:
      type: http
      scheme: bearer
      description: "The same key, sent as `Authorization: Bearer <key>`."
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Unknown job or report.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Unauthorized:
      description: The API key is missing or not accepted.
      headers:
        WWW-Authenticate: { schema: { type: string, const: Bearer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Busy:
      description: |
        Too much work right now: the simulation executor queue is full, or this client's request budget is spent
        (`code` is `busy` or `rate-limited`). Retry after the delay in `Retry-After`.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Seed:
      type: integer
      minimum: 0
      maximum: 4294967295
      description: Root seed (unsigned 32-bit). Omit to have the server pick one; it is echoed back.
...
    JobProgress:
      type: object
      required: [phase, completed, total]
      properties:
        phase: { type: string, enum: [initial-population, generation, campaign] }
        completed: { type: integer }
        total: { type: integer }
    Job:
      type: object
      required: [id, kind, status, createdAt]
      properties:
        id: { type: string }
        kind: { type: string, enum: [optimize, campaign] }
        status: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, 'null'], format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        seed: { $ref: '#/components/schemas/Seed' }
        progress:
          oneOf:
            - $ref: '#/components/schemas/JobProgress'
            - { type: 'null' }
        reportId: { type: [string, 'null'], description: Set when `status` is `succeeded`. }
        warnings:
          type: array
          items: { type: string }
          description: For example, a solo run scored for a party-only role (healer, buffer, controller).
        error:
          oneOf:
            - $ref: '#/components/schemas/Problem'
            - { type: 'null' }
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable machine-readable code, for example `unknown-monster`, `over-capacity`, `unsupported-context`.
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              code: { type: string }